Let's not purge bar and bling records at all.

+++ src/java/com/threerings/msoy/money/server/MoneyTransactionExpirer.java
@@ -66,9 +66,19 @@ public class MoneyTransactionExpirer
      */
     protected void doPurge ()
     {
-        int coins = _repo.deleteOldTransactions(Currency.COINS, COIN_MAX_AGE);
-        int bars = _repo.deleteOldTransactions(Currency.BARS, BAR_MAX_AGE);
-        int bling = _repo.deleteOldTransactions(Currency.BLING, BAR_MAX_AGE);
+        int coins, bars, bling;
+
+        if (COIN_MAX_AGE > 0) {
+            coins = _repo.deleteOldTransactions(Currency.COINS, COIN_MAX_AGE);
+        } else {
+            coins = 0;
+        }
+        if (BAR_MAX_AGE > 0) {
+            bars = _repo.deleteOldTransactions(Currency.BARS, BAR_MAX_AGE);
+            bling = _repo.deleteOldTransactions(Currency.BLING, BAR_MAX_AGE);
+        } else {
+            bars = bling = 0;
+        }
         int exchange = _repo.deleteOldExchangeRecords(EXCHANGE_MAX_AGE);
         if (coins > 0 || bars > 0 || bling > 0 || exchange > 0) {
             log.info("Removed old money transaction records.",
@@ -84,7 +94,7 @@ public class MoneyTransactionExpirer
 
     protected static final long DAY = 24L * 60 * 60 * 1000L; // the length of 99.4% of days
     protected static final long COIN_MAX_AGE = 10L * DAY; // 10 days
-    protected static final long BAR_MAX_AGE = 365L * DAY; // approx 1 year
+    protected static final long BAR_MAX_AGE = -1L; // never expire
     protected static final long EXCHANGE_MAX_AGE = 30L * DAY; // 30 days
 
     protected static final long PURGE_INTERVAL = 60 * 60 * 1000; // 1 hour
